Sources/WebServer/requirements.txt
Click==7.0
dlib==19.19.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.17.4
Pillow==6.2.1
